% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kt-package.R
\docType{package}
\name{kt-package}
\alias{kt-package}
\alias{kt}
\title{kt: A package for ...}
\description{
Computational analysis of KT protein biodiversity.
}
\author{
\strong{Maintainer}: Artem Babaian \email{a.babaian@utoronto.ca}

}
