% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_kt_graph.R
\name{plot_kt_graph}
\alias{plot_kt_graph}
\title{Plot KT Graph}
\usage{
plot_kt_graph(
  graph,
  main = "KT Protein Network",
  vertex_label_size = 0.8,
  show_labels = FALSE,
  legend_position = "topright"
)
}
\arguments{
\item{graph}{An igraph object with visual properties applied}

\item{main}{Character, title for the plot. Default: "KT Protein Network"}

\item{vertex_label_size}{Numeric, size of vertex labels. Default: 0.8}

\item{show_labels}{Logical, whether to show vertex labels. Default: FALSE}

\item{legend_position}{Character, position for legend if coloring by attribute.
Options: "topleft", "topright", "bottomleft", "bottomright", "none"
Default: "topright"}
}
\value{
Plots the graph (invisible return)
}
\description{
Create a plot of the KT graph using base R graphics
}
\examples{
\dontrun{
g <- create_kt_graph(hq_alignments)
g <- apply_kt_graph_visuals(g, node_color_by = "family")
plot_kt_graph(g, show_labels = TRUE)
}

}
