% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_kt_graph.R
\name{plot_kt_graph_summary}
\alias{plot_kt_graph_summary}
\title{Create KT Graph Summary Plot}
\usage{
plot_kt_graph_summary(graph, stats_list)
}
\arguments{
\item{graph}{An igraph object from create_kt_graph()}

\item{stats_list}{Output from get_kt_graph_stats()}
}
\value{
Creates a multi-panel plot (invisible return)
}
\description{
Creates a multi-panel summary plot showing network statistics
}
\examples{
\dontrun{
g <- create_kt_graph(hq_alignments)
stats <- get_kt_graph_stats(g)
plot_kt_graph_summary(g, stats)
}

}
